Define enum in the extra ExecAPI spec for JobState
